Support independent paths
