+++ index.js
@@ -1,5 +1,7 @@
 import Context from './context';
 import Element from './element';
+import Path2D from './path2d';
 
 export {Context};
 export {Element};
+export {Path2D};
