+++ test/rendering.test.js
@@ -21,6 +21,7 @@ import text from './tests/text'
 import tiger from './tests/tiger'
 import transform from './tests/transform'
 import pattern from "./tests/pattern";
+import path2D from './tests/path2D';
 
 const tests = {
     tiger,
@@ -43,7 +44,8 @@ const tests = {
     setLineDash,
     text,
     transform,
-    pattern
+    pattern,
+    path2D
 };
 
 const config = {
