+++ utils.js
@@ -16,4 +16,14 @@ function debug(...data) {
     }
 }
 
-export {toString, debug};
+
+//helper function to format a string
+function format(str, args) {
+    var keys = Object.keys(args), i;
+    for (i=0; i<keys.length; i++) {
+        str = str.replace(new RegExp("\\{" + keys[i] + "\\}", "gi"), args[keys[i]]);
+    }
+    return str;
+}
+
+export {toString, debug, format};
